chore: reorganize components for better structure
