+++ src/App.tsx
@@ -1,7 +1,7 @@
 import './styles/output-stars.css';
 import { Outlet } from 'react-router-dom';
 import './App.css';
-import Navbar from './components/Navbar';
+import Navbar from './components/ui/Navbar';
 import { Analytics } from '@vercel/analytics/react';
 
 function App() {
